# api/server/documentation/api_docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Jamhurican API Documentation",
    "description": "This is official documentaion of Jamhurican application developed by Francis Mwangi.",
    "contact": {
      "email": "mwangif184@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1/"
    },

    {
      "url": "https://jamhurican.com/api/v1/"
    }
  ],
...
  "paths": {
    "/users": {
      "post": {
        "tags": ["Users"],
        "summary": "Register normal user account",
        "operationId": "createUser",
        "requestBody": {
          "description": "User object to be added",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created successully",
            "content": {}
          },
          "400": {
            "description": "Invalid user details supplied",
            "content": {}
          },
          "504": {
            "description": "Server error occurred",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
    "/users/": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "getUsers",
        "parameters": [],
        "responses": {
          "404": {
            "description": "No users found",
            "content": {}
          },
          "200": {
            "description": "Users retrieved successfully"
          },
          "401": {
            "description": "Unauthorized request, please login"
          },
          "504": {
            "description": "Server error occurred",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },

          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "isAdmin": {
            "type": "string"
          }
        },
        "xml": {
          "name": "User"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}